realtime-sample/requirements.txt
python-dotenv

websockets
fastapi[standard]
beautifulsoup4
aiohttp

azure-cognitiveservices-speech